% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_10_cm_to_international.R
\name{icd_10_cm_to_international}
\alias{icd_10_cm_to_international}
\title{Map ICD-10-CM to International ICD-10 (WHO)}
\usage{
icd_10_cm_to_international(icd_10_cm_codes)
}
\arguments{
\item{icd_10_cm_codes}{Character vector of ICD-10-CM codes to convert.}
}
\value{
A character vector of mapped codes (same length and order as input).
}
\description{
Converts a vector of ICD-10-CM codes to their closest International ICD-10 parent code.
It relies on the package data object \code{icd_10_international_codes} as the reference.
}
