% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_10_international_codes}
\alias{icd_10_international_codes}
\title{A dataset containing all ICD-10 international codes}
\format{
icd-10 international codes
}
\usage{
icd_10_international_codes
}
\description{
A dataset containing all ICD-10 international codes
}
\keyword{datasets}
