% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supported_icd_10_codes}
\alias{supported_icd_10_codes}
\title{A dataset containing all supported trauma ICD codes and the one-hot encoded index in the encoder}
\format{
icd-10 codes with one hot encoding index
}
\usage{
supported_icd_10_codes
}
\description{
A dataset containing all supported trauma ICD codes and the one-hot encoded index in the encoder
}
\keyword{datasets}
